Calculate total curvature from polynomial functions fitted to landmarked specimens.

Calculates total curvature along bounded polynomials fitted to curved specimens. Polynomials are fitted to landmarked specimens using package Momocs.

curvature_poly(Momocs_poly, x_range, subdiv = 1000, param = FALSE)

Arguments

Momocs_poly
a polynomial list object generated by Momocs:::npoly.default() or Momocs:::opoly.default(). This function cannot yet handle OpnCoe objects created by e.g. Momocs:::npoly.Opn().

x_range
the lower and upper x-value bounds to calculate curvature. These are named 'baselines' in Momocs polynomial objects.

subdiv
number of subdivisions made within x_range. Curvature is calculated at every subdivision and summed to compute total curvature. Default is 1000. The same number of subdivisions will be applied to all curves, regardless of arclength.

param
whether to parameterize the curve by arc length (seepracma::arclength). Setting param=TRUE greatly increases computing time.

Value

a numeric representing total curvature. Units are radians.

Examples

library(Momocs)

# x <- 1:10
# y <- x^2
#
# # a fictional landmark dataset
# mdat <- matrix(c(x,y), nrow=10, ncol=2)
#
# # fit second-order polynomial to landmarks
# my_poly <- Momocs::npoly(mdat, 2)

# extract x bounds from Momocs polynomial object
x_range <- c(my_poly$baseline1[[1]], my_poly$baseline2[[1]])
#> Error in eval(expr, envir, enclos) object 'my_poly' not found

# compute point-wise curvature 100 times
curvature_poly(my_poly, x_range, 100)
#> Error in stopifnot(is.atomic(x_range)) object 'x_range' not found