Calculate total curvature from polynomial functions fitted to landmarked specimens.

Calculates point-wise curvature along bounded polynomials fitted to curved specimens. Polynomials are fitted to landmarked specimens using package Momocs.

total_curvature(Momocs_poly, x_range, subdiv)

Arguments

Momocs_poly
a polynomial list object generated by Momocs:::npoly.default() or Momocs:::opoly.default(). This function cannot yet handle OpnCoe objects created by e.g. Momocs:::npoly.Opn().

x_range
the lower and upper x-values bounds to calculate curvature. These are named 'baselines' in Momocs polynomial objects.

subdiv
number of subdivisions made within x_range. Curvature is calculated at every subdivision and summed to compute total curvature. Default is 1000. The same number of subdivisions will be applied to all curves, regardless of arclength.

Value

a list. The first element contains point-wise curvature. Units are degrees per unit length. These values are useful for inspecting local curvature features. The second element is total curvature. Units are degrees. Since arclength (ie specimen size) does not affect the number of times point-wise curvature is computed (see subdiv), there is no need to divide total curvature by arclength to adjust for specimen size.

Examples

library(Momocs)

x <- 1:10
y <- x^2

# a fictional landmark dataset
mdat <- matrix(c(x,y), nrow=10, ncol=2)

# fit second-order polynomial to landmarks
my_poly <- Momocs::npoly(mdat, 2)

#> Warning: essentially perfect fit: summary may be unreliable

# extract x bounds from Momocs polynomial object
x_range <- c(my_poly$baseline1[[1]], my_poly$baseline2[[1]])

# compute point-wise curvature 100 times
total_curvature(my_poly, x_range, 100)

#> $k
#>                  x
#>   [1,] 10.24938063
#>   [2,]  4.48808106
#>   [3,]  2.60601776
#>   [4,]  1.74146781
#>   [5,]  1.26496836
#>   [6,]  0.97086608
#>   [7,]  0.77478439
#>   [8,]  0.63652130
#>   [9,]  0.53484910
#>  [10,]  0.45749974
#>  [11,]  0.39708098
#>  [12,]  0.34883581
#>  [13,]  0.30958516
#>  [14,]  0.27717806
#>  [15,]  0.25002806
#>  [16,]  0.22702607
#>  [17,]  0.20733766
#>  [18,]  0.19033218
#>  [19,]  0.17552513
#>  [20,]  0.16253883
#>  [21,]  0.15107503
#>  [22,]  0.14089541
#>  [23,]  0.13180749
#>  [24,]  0.12365434
#>  [25,]  0.11630684
#>  [26,]  0.10965794
#>  [27,]  0.10361815
#>  [28,]  0.09811222
#>  [29,]  0.09307641
#>  [30,]  0.08845645
#>  [31,]  0.08420585
#>  [32,]  0.08028459
#>  [33,]  0.07665806
#>  [34,]  0.07329629
#>  [35,]  0.07017320
#>  [36,]  0.06726541
#>  [37,]  0.06455274
#>  [38,]  0.06201551
#>  [39,]  0.05964172
#>  [40,]  0.05741539
#>  [41,]  0.05532403
#>  [42,]  0.05335646
#>  [43,]  0.05150267
#>  [44,]  0.04975368
#>  [45,]  0.04810137
#>  [46,]  0.04653843
#>  [47,]  0.04505824
#>  [48,]  0.04365481
#>  [49,]  0.04232267
#>  [50,]  0.04105686
#>  [51,]  0.03985289
#>  [52,]  0.03870654
#>  [53,]  0.03761407
#>  [54,]  0.03657198
#>  [55,]  0.03557710
#>  [56,]  0.03462649
#>  [57,]  0.03371743
#>  [58,]  0.03284743
#>  [59,]  0.03201417
#>  [60,]  0.03121554
#>  [61,]  0.03044944
#>  [62,]  0.02971414
#>  [63,]  0.02900796
#>  [64,]  0.02832930
#>  [65,]  0.02767732
#>  [66,]  0.02704935
#>  [67,]  0.02644476
#>  [68,]  0.02586236
#>  [69,]  0.02530101
#>  [70,]  0.02475967
#>  [71,]  0.02423735
#>  [72,]  0.02373313
#>  [73,]  0.02324615
#>  [74,]  0.02277558
#>  [75,]  0.02232067
#>  [76,]  0.02188070
#>  [77,]  0.02145498
#>  [78,]  0.02104289
#>  [79,]  0.02064381
#>  [80,]  0.02025718
#>  [81,]  0.01988246
#>  [82,]  0.01951915
#>  [83,]  0.01916677
#>  [84,]  0.01882485
#>  [85,]  0.01849298
#>  [86,]  0.01817073
#>  [87,]  0.01785782
#>  [88,]  0.01755377
#>  [89,]  0.01725790
#>  [90,]  0.01697055
#>  [91,]  0.01669108
#>  [92,]  0.01641918
#>  [93,]  0.01615458
#>  [94,]  0.01589700
#>  [95,]  0.01564618
#>  [96,]  0.01540189
#>  [97,]  0.01516388
#>  [98,]  0.01493194
#>  [99,]  0.01470584
#> [100,]  0.01448540
#> [101,]  0.01427040
#> 
#> $total_k
#> [1] 0.2980114
#>